Easing completely rebuilt
Easing works now differently. The calculation of steps is made in the central class crawly
and the calculation for the actual angle is done in real time instead of
beforehand. This makes it more flexible. A lot of methods with different version remain.
This has to be clear up later.

Signed-off-by: Marcus Jacobs <west.wallaby@posteo.de>

next attempt

No idea